refactor: split auth.rs and handler.rs into modules
Split auth.rs (824 lines) into auth/ directory:
- mod.rs: module entry with public re-exports
- config.rs: OAuthConfig, ApiKeyConfig, AuthConfig
- manager.rs: AuthManager implementation
- token.rs: TokenStore, TokenInfo
- types.rs: OAuthProvider, AuthProvider, GeneratedApiKey, AuthContext
- tests.rs: unit tests

Split handler.rs (728 lines) into handler/ directory:
- mod.rs: module entry with public re-exports
- types.rs: ToolExecutionResult
- config.rs: HandlerConfig
- core.rs: HandlerCore implementation
- standard.rs: CratesDocsHandler
- core_handler.rs: CratesDocsHandlerCore

Benefits:
- Each file < 300 lines, easier to read and maintain
- Better separation of concerns
- Maintains backward compatibility via pub use re-exports
- All CI checks pass
