+++ src/server/auth/config.rs
@@ -1,25 +1,4 @@
-//! Authentication module
-//!
-//! Provides OAuth 2.0 and API Key authentication support.
-//!
-//! # Authentication Methods
-//!
-//! - **OAuth 2.0**: Full OAuth flow with GitHub, Google, Keycloak support
-//! - **API Key**: Simple API key authentication with secure hashing
-//!
-//! # Example
-//!
-//! ```rust,no_run
-//! use crates_docs::server::auth::{ApiKeyConfig, AuthConfig};
-//!
-//! // Create API Key configuration
-//! let api_key_config = ApiKeyConfig {
-//!     enabled: true,
-//!     keys: vec!["sk_live_xxx".to_string()],
-//!     header_name: "X-API-Key".to_string(),
-//!     ..Default::default()
-//! };
-//! ```
+//! Authentication configuration
 
 use crate::error::{Error, Result};
 use url::Url;
@@ -29,6 +8,8 @@ use api_keys_simplified::{
     ApiKey, ApiKeyManagerV0, Environment, ExposeSecret, HashConfig, KeyConfig, KeyStatus,
     SecureString,
 };
+
+use super::types::{GeneratedApiKey, OAuthProvider};
 
 /// OAuth configuration
 #[derive(Debug, Clone, serde::Deserialize, serde::Serialize)]
@@ -49,19 +30,6 @@ pub struct OAuthConfig {
     pub scopes: Vec<String>,
     /// Authentication provider type
     pub provider: OAuthProvider,
-}
-
-/// OAuth provider type
-#[derive(Debug, Clone, serde::Deserialize, serde::Serialize, PartialEq)]
-pub enum OAuthProvider {
-    /// Custom OAuth provider
-    Custom,
-    /// GitHub OAuth
-    GitHub,
-    /// Google OAuth
-    Google,
-    /// Keycloak
-    Keycloak,
 }
 
 impl Default for OAuthConfig {
@@ -212,21 +180,6 @@ impl OAuthConfig {
     pub fn to_mcp_config(&self) -> Result<()> {
         Err(Error::config("oauth", "feature is not enabled"))
     }
-}
-
-/// API Key generation result.
-///
-/// The plain-text key should be shown once to the operator and stored securely.
-/// The hash should be persisted in configuration or external secret storage.
-#[derive(Debug, Clone, serde::Deserialize, serde::Serialize)]
-#[cfg(feature = "api-key")]
-pub struct GeneratedApiKey {
-    /// Plain-text API key for one-time display
-    pub key: String,
-    /// Stable key identifier derived from the key
-    pub key_id: String,
-    /// Argon2 PHC hash to store and verify against
-    pub hash: String,
 }
 
 /// API Key configuration
@@ -465,360 +418,5 @@ impl AuthConfig {
     #[cfg(not(feature = "api-key"))]
     pub fn is_enabled(&self) -> bool {
         self.oauth.enabled
-    }
-}
-
-/// Authentication provider type
-#[derive(Debug, Clone, Copy, PartialEq, Eq)]
-pub enum AuthProvider {
-    /// No authentication
-    None,
-    /// OAuth authentication
-    OAuth,
-    /// API Key authentication
-    #[cfg(feature = "api-key")]
-    ApiKey,
-}
-
-/// Authentication context
-#[derive(Debug, Clone)]
-pub struct AuthContext {
-    /// Authentication provider used
-    pub provider: AuthProvider,
-    /// User ID (if available)
-    pub user_id: Option<String>,
-    /// User email (if available)
-    pub user_email: Option<String>,
-    /// API key identifier (if API key auth)
-    #[cfg(feature = "api-key")]
-    pub api_key_id: Option<String>,
-}
-
-impl AuthContext {
-    /// Create a new authentication context
-    #[must_use]
-    pub fn new(provider: AuthProvider) -> Self {
-        Self {
-            provider,
-            user_id: None,
-            user_email: None,
-            #[cfg(feature = "api-key")]
-            api_key_id: None,
-        }
-    }
-
-    /// Check if authentication is authenticated
-    #[must_use]
-    pub fn is_authenticated(&self) -> bool {
-        !matches!(self.provider, AuthProvider::None)
-    }
-}
-
-/// Authentication manager
-#[derive(Default)]
-pub struct AuthManager {
-    config: OAuthConfig,
-    #[cfg(feature = "api-key")]
-    api_key_config: ApiKeyConfig,
-}
-
-impl AuthManager {
-    /// Create a new authentication manager
-    pub fn new(config: OAuthConfig) -> Result<Self> {
-        config.validate()?;
-        Ok(Self {
-            config,
-            #[cfg(feature = "api-key")]
-            api_key_config: ApiKeyConfig::default(),
-        })
-    }
-
-    /// Create a new authentication manager with full config
-    #[cfg(feature = "api-key")]
-    pub fn with_config(config: AuthConfig) -> Result<Self> {
-        config.validate()?;
-        Ok(Self {
-            config: config.oauth,
-            api_key_config: config.api_key,
-        })
-    }
-
-    /// Check if authentication is enabled
-    #[must_use]
-    #[cfg(feature = "api-key")]
-    pub fn is_enabled(&self) -> bool {
-        self.config.enabled || self.api_key_config.enabled
-    }
-
-    /// Check if authentication is enabled
-    #[must_use]
-    #[cfg(not(feature = "api-key"))]
-    pub fn is_enabled(&self) -> bool {
-        self.config.enabled
-    }
-
-    /// Get OAuth configuration
-    #[must_use]
-    pub fn config(&self) -> &OAuthConfig {
-        &self.config
-    }
-
-    /// Get API key configuration
-    #[cfg(feature = "api-key")]
-    #[must_use]
-    pub fn api_key_config(&self) -> &ApiKeyConfig {
-        &self.api_key_config
-    }
-
-    /// Validate API key
-    #[cfg(feature = "api-key")]
-    #[must_use]
-    pub fn validate_api_key(&self, key: &str) -> bool {
-        self.api_key_config.is_valid_key(key)
-    }
-
-    /// Generate a new API key and hash pair
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if key generation fails
-    #[cfg(feature = "api-key")]
-    pub fn generate_api_key(&self) -> Result<GeneratedApiKey> {
-        self.api_key_config.generate_key()
-    }
-
-    /// Extract API key from request headers
-    #[cfg(feature = "api-key")]
-    #[must_use]
-    pub fn extract_api_key_from_headers(
-        &self,
-        headers: &std::collections::HashMap<String, String>,
-    ) -> Option<String> {
-        headers.get(&self.api_key_config.header_name).cloned()
-    }
-}
-
-/// Simple in-memory token store (production should use Redis or database)
-#[derive(Default)]
-pub struct TokenStore {
-    tokens: std::sync::RwLock<std::collections::HashMap<String, TokenInfo>>,
-}
-
-/// OAuth token information
-#[derive(Debug, Clone, serde::Deserialize, serde::Serialize)]
-pub struct TokenInfo {
-    /// Access token
-    pub access_token: String,
-    /// Refresh token (optional)
-    pub refresh_token: Option<String>,
-    /// Token expiration time
-    pub expires_at: chrono::DateTime<chrono::Utc>,
-    /// Authorization scopes
-    pub scopes: Vec<String>,
-    /// User ID (optional)
-    pub user_id: Option<String>,
-    /// User email (optional)
-    pub user_email: Option<String>,
-}
-
-impl TokenStore {
-    /// Create a new token store
-    #[must_use]
-    pub fn new() -> Self {
-        Self::default()
-    }
-
-    /// Store token
-    pub fn store_token(&self, key: String, token: TokenInfo) {
-        let mut tokens = self.tokens.write().unwrap();
-        tokens.insert(key, token);
-    }
-
-    /// Get token
-    pub fn get_token(&self, key: &str) -> Option<TokenInfo> {
-        let tokens = self.tokens.read().unwrap();
-        tokens.get(key).cloned()
-    }
-
-    /// Remove token
-    pub fn remove_token(&self, key: &str) {
-        let mut tokens = self.tokens.write().unwrap();
-        tokens.remove(key);
-    }
-
-    /// Cleanup expired tokens
-    pub fn cleanup_expired(&self) {
-        let now = chrono::Utc::now();
-        let mut tokens = self.tokens.write().unwrap();
-        tokens.retain(|_, token| token.expires_at > now);
-    }
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn test_oauth_config_default() {
-        let config = OAuthConfig::default();
-        assert!(!config.enabled);
-        assert!(config.client_id.is_none());
-    }
-
-    #[test]
-    fn test_oauth_config_github() {
-        let config = OAuthConfig::github(
-            "client_id".to_string(),
-            "client_secret".to_string(),
-            "http://localhost:8080/callback".to_string(),
-        );
-        assert!(config.enabled);
-        assert_eq!(config.provider, OAuthProvider::GitHub);
-    }
-
-    #[test]
-    fn test_oauth_config_validate() {
-        let config = OAuthConfig::default();
-        assert!(config.validate().is_ok());
-
-        let config = OAuthConfig {
-            enabled: true,
-            ..Default::default()
-        };
-        assert!(config.validate().is_err());
-    }
-
-    #[cfg(feature = "api-key")]
-    #[test]
-    fn test_api_key_config_default() {
-        let config = ApiKeyConfig::default();
-        assert!(!config.enabled);
-        assert!(config.keys.is_empty());
-        assert_eq!(config.header_name, "X-API-Key");
-        assert_eq!(config.key_prefix, "sk");
-    }
-
-    #[cfg(feature = "api-key")]
-    #[test]
-    fn test_api_key_config_validate() {
-        let config = ApiKeyConfig::default();
-        assert!(config.validate().is_ok());
-
-        let config = ApiKeyConfig {
-            enabled: true,
-            header_name: String::new(),
-            ..Default::default()
-        };
-        assert!(config.validate().is_err());
-    }
-
-    #[cfg(feature = "api-key")]
-    #[test]
-    fn test_api_key_is_valid() {
-        let key_config = ApiKeyConfig {
-            enabled: true,
-            ..Default::default()
-        };
-        let generated = key_config.generate_key().unwrap();
-
-        let config = ApiKeyConfig {
-            enabled: true,
-            keys: vec![generated.hash.clone()],
-            ..Default::default()
-        };
-
-        assert!(config.is_valid_key(&generated.key));
-        assert!(!config.is_valid_key("invalid_key"));
-    }
-
-    #[cfg(feature = "api-key")]
-    #[test]
-    fn test_api_key_disabled_allows_all() {
-        let config = ApiKeyConfig::default();
-        assert!(!config.enabled);
-
-        // When disabled, all keys should be accepted
-        assert!(config.is_valid_key("any_key"));
-    }
-
-    #[cfg(feature = "api-key")]
-    #[test]
-    fn test_api_key_plaintext_fallback() {
-        let config = ApiKeyConfig {
-            enabled: true,
-            keys: vec!["legacy_plaintext_key".to_string()],
-            ..Default::default()
-        };
-
-        assert!(config.is_valid_key("legacy_plaintext_key"));
-        assert!(!config.is_valid_key("legacy_plaintext_key_2"));
-    }
-
-    #[cfg(feature = "api-key")]
-    #[test]
-    fn test_api_key_legacy_hashed_verification() {
-        let config = ApiKeyConfig::default();
-        let legacy_hash = config
-            .normalize_key_material("legacy_plaintext_key")
-            .unwrap();
-
-        let enabled_config = ApiKeyConfig {
-            enabled: true,
-            keys: vec![legacy_hash],
-            ..Default::default()
-        };
-
-        assert!(enabled_config.is_valid_key("legacy_plaintext_key"));
-        assert!(!enabled_config.is_valid_key("legacy_plaintext_key_2"));
-    }
-
-    #[cfg(feature = "api-key")]
-    #[test]
-    fn test_api_key_generate_key_returns_hash_and_key() {
-        let config = ApiKeyConfig::default();
-        let generated = config.generate_key().unwrap();
-
-        assert!(
-            generated.key.starts_with("sk-")
-                || generated.key.starts_with("sk_")
-                || generated.key.starts_with("sk")
-        );
-        assert!(!generated.key_id.is_empty());
-        assert!(generated.hash.starts_with("$argon2"));
-    }
-
-    #[test]
-    fn test_auth_config_default() {
-        let config = AuthConfig::default();
-        assert!(!config.is_enabled());
-    }
-
-    #[test]
-    fn test_auth_context() {
-        let ctx = AuthContext::new(AuthProvider::None);
-        assert!(!ctx.is_authenticated());
-
-        let ctx = AuthContext::new(AuthProvider::OAuth);
-        assert!(ctx.is_authenticated());
-    }
-
-    #[test]
-    fn test_token_store() {
-        let store = TokenStore::new();
-        let token = TokenInfo {
-            access_token: "test_token".to_string(),
-            refresh_token: None,
-            expires_at: chrono::Utc::now() + chrono::Duration::hours(1),
-            scopes: vec!["read".to_string()],
-            user_id: None,
-            user_email: None,
-        };
-
-        store.store_token("key".to_string(), token.clone());
-        assert!(store.get_token("key").is_some());
-        assert!(store.get_token("nonexistent").is_none());
-
-        store.remove_token("key");
-        assert!(store.get_token("key").is_none());
     }
 }
